backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
together>=1.0.0
google-api-python-client>=2.110.0
python-dotenv>=1.0.0
